+++ test/fixtures/d.js
@@ -2,6 +2,7 @@
 /**
  * Parse tag type string "{Array|Object}" etc.
  *
+ * @name is arbitrary
  * @param {String} str
  * @return {Array}
  * @api public