+++ History.md
@@ -1,4 +1,15 @@
 
+0.2.0 / 2012-02-23 
+==================
+
+  * Added `-r, --raw` support. Closes #48
+
+0.1.3 / 2011-12-08 
+==================
+
+  * Added: allow arbitrary tags [logicalparadox]
+  * Fixed function whitespace [TooTallNate]
+
 0.1.2 / 2011-10-22 
 ==================